Use detached threads instead of forking for reconstruction
Switch to thread-pool from libfork, so that we can detach from the running thread and
push all buildings across all tiles onto the
queue.
The threads pick the buildings off the queue one-by-one and the reconstruction does not need
to wait for the whole tile to finish in order to
continue with buildings from the next tile.